redo the reca
If you bought the Codex template (based on your issue description I believe you did) the issue is the template. You need to comment out lines 620-661 in the main.js file. The button is expecting a response but doesn't get it because you're not using the php it was originally setup to work with.

But also double check your Scripts section at the bottom of index.html the scripts should be src="/js and not src"js.